requirements.txt
asgiref==3.11.0
Django==6.0.1
djangorestframework==3.16.1
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.11
sqlparse==0.5.5
whitenoise==6.11.0
